product_management_service/requirements.txt
SQLAlchemy~=2.0.23
fastapi~=0.105.0
PyYAML~=6.0.1
uvicorn~=0.24.0.post1
pydantic~=2.5.2
python-dotenv~=1.0.0
alembic~=1.13.0
psycopg2~=2.9.9
asyncpg~=0.29.0
python-multipart~=0.0.6
aiohttp~=3.9.1
pydantic[email]